Changed play() function to fit rubric. Added a lot of comments.
play() now takes Hand& and Deck& as parameters, and is called directly on a Card object itself. Card object is accessed ??? some way during play - currently another function called getCard is made that gets the card from the hand via an index. It returns an order that player will put in player list.